WindSpeedExploration Octave Code
This 'WindSpeedExploration.m' file is code for the Octave software which I used for an experiment to determine the most optimal motor load, which turned out to be 75%.

The code uses the three files below them which have no extension (could be .txt), containing the values for the lift force of each dataset.

The code takes the preferrable range of values to exclude 0 values (which is done by making a 'b' variable, which only includes a portion of the initial loaded file) and takes the mean, standard deviation, and SNR (signal-to-noise ratio) for each, which is shown as a "text message" (I'm not sure exactly what to call it) output which allows the user to compare the values of these 3 runs at different motor loads.

The message summarizing the data processing is only shown at the end, but the code also creates 3 figures:
-3 subplots with the raw data
-All 3 datasets with mean lines on a single plot, color-coded.
-The histograms of all 3 datasets, showing how many times a certain value was recorded.

If using this program for your own purposes, note that you will have to change the data files to your preference and change the 'b' variables to your desired range.